Move error definition in a more suited file
