+++ src/GameState.h
@@ -4,17 +4,6 @@
 #include "JsonFwd.h"
 #include <string>
 
-class Game;
-
-struct CannotSaveCurrentGameState {};
-struct CannotSaveInHyperspace : public CannotSaveCurrentGameState {};
-struct CannotSaveDeadPlayer : public CannotSaveCurrentGameState {};
-struct InvalidGameStartLocation {
-	std::string error;
-	InvalidGameStartLocation(const std::string &error_) :
-		error(error_) {}
-};
-
 class GameState
 {
 	public:
