+++ src/Pi.cpp
@@ -17,7 +17,6 @@
 #include "GameConfig.h"
 #include "GameConfSingleton.h"
 #include "GameLog.h"
-#include "GameSaveError.h"
 #include "GameState.h"
 #include "InGameViews.h"
 #include "Intro.h"
